fix descriptions, convert apps to snap if possible
In order to make sure that the functionality of the script is operating at its best, I have sought to simplify the script by using the relatively easy (and consistently functional) snapd package manager for as many of the programs as possible. In the same spirit, I have removed all unnecessary repos (preferring apps with packages in the Ubuntu repos). Additionally, the descriptions are somewhat fixed, as in they are more descriptive of the programs.